python scripts and raw data for backtesting/requirements.txt
backtesting~=0.3.3
stockstats~=0.4.1
scikit-learn~=1.1.1
tensorflow~=2.9.1
tensorflow-text
torch
transformers
keras
pandas
numpy
emoji
tweet-preprocessor
xgboost
statsmodels